requirements.txt
Flask==3.0.0
PyMongo==4.6.0
bcrypt==4.1.2
python-dotenv==1.0.0
gunicorn==21.2.0
Flask-Talisman==1.1.0
Werkzeug==3.0.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Development (optional)
black==23.11.0
flake8==6.1.0
